pyproject.toml
[tool.poetry]
name = "carnet-ai-recognition-cli"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "carnet_ai_recognition_cli"}]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.26.142"
argparse = "^1.4.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.4"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
